refactor: del opp generatoren for arkitekturgrafen
269 linjer i én fil er blitt en CLI på 55 og fire deler med ett ansvar
hver: parse, resolve, scan, serialize.

Grunnen den var én fil er borte nå. Node krever .ts i import-stien, tsc
forbød den, og allowImportingTsExtensions er det ene flagget som lar
begge lese samme sti. Det krever noEmit, som repoet har, og Next-bygget
godtar det uten å skrive om tsconfig.
